requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiocontextvars==0.2.2     # via -r requirements.in
aiodns==2.0.0             # via faust
aiohttp-cors==0.7.0       # via faust
aiohttp==3.6.2            # via aiohttp-cors, faust
async-timeout==3.0.1      # via aiohttp
atomicwrites==1.3.0       # via -r requirements.in
attrs==19.3.0             # via aiohttp
blis==0.4.1               # via spacy, thinc
catalogue==1.0.0          # via spacy
cchardet==2.1.5           # via faust
certifi==2019.11.28       # via requests
cffi==1.14.0              # via pycares
chardet==3.0.4            # via aiohttp, requests
ciso8601==2.1.3           # via faust
click==7.0                # via faust
colorclass==2.2.0         # via faust
colorlog==4.1.0           # via mode
croniter==0.3.31          # via faust
cymem==2.0.3              # via preshed, spacy, thinc
cython==0.29.15           # via -r requirements.in, faust
faust[fast]==1.10.4       # via -r requirements.in
idna==2.9                 # via requests, yarl
importlib-metadata==1.5.0  # via catalogue
kafka-python==1.4.7       # via robinhood-aiokafka
mode==4.3.2               # via faust
multidict==4.7.5          # via aiohttp, yarl
murmurhash==1.0.2         # via preshed, spacy, thinc
mypy-extensions==0.4.3    # via faust, mode
numpy==1.18.1             # via blis, spacy, thinc
opentracing==1.3.0        # via faust
orjson==2.5.1             # via faust
plac==1.1.3               # via spacy, thinc
preshed==3.0.2            # via spacy, thinc
py==1.8.1                 # via -r requirements.in
pycares==3.1.1            # via aiodns
pycparser==2.19           # via cffi
pymongo==3.10.1           # via -r requirements.in
python-dateutil==2.8.1    # via croniter
python-dotenv==0.12.0     # via -r requirements.in
python-rocksdb==0.7.0     # via -r requirements.in
requests==2.23.0          # via spacy
robinhood-aiokafka==1.1.6  # via faust
setproctitle==1.1.10      # via faust
six==1.14.0               # via python-dateutil
spacy==2.2.3              # via -r requirements.in
srsly==1.0.1              # via spacy, thinc
terminaltables==3.1.0     # via faust
thinc==7.3.1              # via spacy
tqdm==4.43.0              # via thinc
typing-extensions==3.7.4.1  # via mode
urllib3==1.25.8           # via requests
vadersentiment==3.2.1     # via -r requirements.in
venusian==1.2.0           # via faust
wasabi==0.6.0             # via spacy, thinc
yarl==1.4.2               # via aiohttp, faust
zipp==3.0.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
